+++ Octopath-Traveler-Model/Entities/SkillInfo.cs
@@ -5,6 +5,6 @@ public class SkillInfo
     public string Name { get; set; }
     public DamageType Type { get; set; }
     public string Description { get; set; }
-    public double Modifier { get; set; }
+    public decimal Modifier { get; set; }
     public SkillTarget Target { get; set; }
 }
