+++ Octopath-Traveler-Model/JsonData/BeastSkillJsonData.cs
@@ -3,7 +3,7 @@ namespace Octopath_Traveler_Model;
 public class BeastSkillJsonData
 {
     public string Name { get; set; }
-    public float Modifier { get; set; }
+    public decimal Modifier { get; set; }
     public string Description { get; set; }
     public string Target { get; set; }
     public int Hits { get; set; }
