+++ Octopath-Traveler-Model/JsonData/TravelerSkillJsonData.cs
@@ -7,6 +7,6 @@ public class TravelerSkillJsonData
     public string Type { get; set; }
     public string Description { get; set; }
     public string Target { get; set; }
-    public float Modifier { get; set; }
+    public decimal Modifier { get; set; }
     public string Boost { get; set; }
 }
